+++ app/lib/placeholder-data.js
@@ -1,0 +1,267 @@
+const questionArr = [
+    {
+        questionText: '1Which of the following options would you choose?',
+        options: [
+            {
+                label: 'This is A',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is B',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is C',
+                isCorrectAnswer: true
+            },
+            {
+                label: 'This is D',
+                isCorrectAnswer: false
+            }
+        ],
+        correctAnswer: {
+            label: 'This is C',
+            isCorrectAnswer: true
+        },
+        explanation: 'This is how it is'
+    },
+    {
+        questionText: '2Which of the following options would you choose?',
+        options: [
+            {
+                label: 'This is A',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is B',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is C',
+                isCorrectAnswer: true
+            },
+            {
+                label: 'This is D',
+                isCorrectAnswer: false
+            }
+        ],
+        correctAnswer: {
+            label: 'This is C',
+            isCorrectAnswer: true
+        },
+        explanation: 'This is how it is'
+    },
+    {
+        questionText: '3Which of the following options would you choose?',
+        options: [
+            {
+                label: 'This is A',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is B',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is C',
+                isCorrectAnswer: true
+            },
+            {
+                label: 'This is D',
+                isCorrectAnswer: false
+            }
+        ],
+        correctAnswer: {
+            label: 'This is C',
+            isCorrectAnswer: true
+        },
+        explanation: 'This is how it is'
+    },
+    {
+        questionText: '4Which of the following options would you choose?',
+        options: [
+            {
+                label: 'This is A',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is B',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is C',
+                isCorrectAnswer: true
+            },
+            {
+                label: 'This is D',
+                isCorrectAnswer: false
+            }
+        ],
+        correctAnswer: {
+            label: 'This is C',
+            isCorrectAnswer: true
+        },
+        explanation: 'This is how it is'
+    },
+    {
+        questionText: '5Which of the following options would you choose?',
+        options: [
+            {
+                label: 'This is A',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is B',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is C',
+                isCorrectAnswer: true
+            },
+            {
+                label: 'This is D',
+                isCorrectAnswer: false
+            }
+        ],
+        correctAnswer: {
+            label: 'This is C',
+            isCorrectAnswer: true
+        },
+        explanation: 'This is how it is'
+    },
+    {
+        questionText: '6Which of the following options would you choose?',
+        options: [
+            {
+                label: 'This is A',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is B',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is C',
+                isCorrectAnswer: true
+            },
+            {
+                label: 'This is D',
+                isCorrectAnswer: false
+            }
+        ],
+        correctAnswer: {
+            label: 'This is C',
+            isCorrectAnswer: true
+        },
+        explanation: 'This is how it is'
+    },
+    {
+        questionText: '7Which of the following options would you choose?',
+        options: [
+            {
+                label: 'This is A',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is B',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is C',
+                isCorrectAnswer: true
+            },
+            {
+                label: 'This is D',
+                isCorrectAnswer: false
+            }
+        ],
+        correctAnswer: {
+            label: 'This is C',
+            isCorrectAnswer: true
+        },
+        explanation: 'This is how it is'
+    },
+    {
+        questionText: '8Which of the following options would you choose?',
+        options: [
+            {
+                label: 'This is A',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is B',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is C',
+                isCorrectAnswer: true
+            },
+            {
+                label: 'This is D',
+                isCorrectAnswer: false
+            }
+        ],
+        correctAnswer: {
+            label: 'This is C',
+            isCorrectAnswer: true
+        },
+        explanation: 'This is how it is'
+    },
+    {
+        questionText: '9Which of the following options would you choose?',
+        options: [
+            {
+                label: 'This is A',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is B',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is C',
+                isCorrectAnswer: true
+            },
+            {
+                label: 'This is D',
+                isCorrectAnswer: false
+            }
+        ],
+        correctAnswer: {
+            label: 'This is C',
+            isCorrectAnswer: true
+        },
+        explanation: 'This is how it is'
+    },
+    {
+        questionText: '10Which of the following options would you choose?',
+        options: [
+            {
+                label: 'This is A',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is B',
+                isCorrectAnswer: false
+            },
+            {
+                label: 'This is C',
+                isCorrectAnswer: true
+            },
+            {
+                label: 'This is D',
+                isCorrectAnswer: false
+            }
+        ],
+        correctAnswer: {
+            label: 'This is C',
+            isCorrectAnswer: true
+        },
+        explanation: 'This is how it is'
+    }
+]
+
+module.exports = {
+    questionArr
+  };
+
